update/refactor: Error less due to file missing

+++ 1 File handle/File handle binary/Deleting record in a binary file.py
@@ -1,17 +1,38 @@
 import pickle
+import os
+import logging
+
+logging.basicConfig(level=logging.INFO, format='%(levelname)s: %(message)s')
 
 
-def bdelete():
+def b_read():
     # Opening a file & loading it
-    with open("studrec.dat", "rb") as F:
-        stud = pickle.load(F)
-        print(stud)
+    if not os.path.exists("student_records.pkl"):
+        logging.warning("File not found")
+        return
 
+    with open("student_records.pkl", "rb") as F:
+        student = pickle.load(F)
+        logging.info("File opened successfully")
+        logging.info("Records in the file are:")
+        for i in student:
+            logging.info(i)
+
+def b_modify():
     # Deleting the Roll no. entered by user
-    rno = int(input("Enter the Roll no. to be deleted: "))
-    with open("studrec.dat", "wb") as F:
-        rec = [i for i in stud if i[0] != rno]
+    if not os.path.exists("student_records.pkl"):
+        logging.warning("File not found")
+        return
+    roll_no = int(input("Enter the Roll No. to be deleted: "))
+    student = 0
+    with open("student_records.pkl", "rb") as F:
+        student = pickle.load(F)
+
+    with open("student_records.pkl", "wb") as F:
+        rec = [i for i in student if i[0] != roll_no]
         pickle.dump(rec, F)
 
 
-bdelete()
+
+b_read()
+b_modify()
